Use metadata for titles and move content out of page files
