+++ bods-nextjs/app/publish/org/[orgId]/dataset/avl/[datasetId]/delete/success/page.tsx
@@ -1,40 +1,15 @@
-'use client';
-
-import Link from 'next/link';
-import { useParams } from 'next/navigation';
-import { useEffect } from 'react';
+import type { Metadata } from 'next';
 import { ProtectedRoute } from '@/components/auth/ProtectedRoute';
+import { AvlDeleteSuccessContent } from '../../_components/AvlDeleteSuccessContent';
 
-function AvlDeleteSuccessPageContent() {
-  const params = useParams();
-  const orgId = params.orgId as string;
-  const listUrl = `/publish/org/${orgId}/dataset/avl`;
-
-  useEffect(() => {
-    document.title = 'Data feed deleted';
-  }, []);
-
-  return (
-    <div className="govuk-width-container">
-      <div className="govuk-main-wrapper">
-        <div className="govuk-grid-row">
-          <div className="govuk-grid-column-full">
-            <h1 className="govuk-heading-xl">Data feed has been deleted</h1>
-            <p className="govuk-body-l">Data feed has been deleted.</p>
-            <Link role="button" className="govuk-button govuk-!-margin-top-5" href={listUrl}>
-              Go back to your data feeds
-            </Link>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
+export const metadata: Metadata = {
+  title: 'Data feed deleted',
+};
 
 export default function AvlDeleteSuccessPage() {
   return (
     <ProtectedRoute>
-      <AvlDeleteSuccessPageContent />
+      <AvlDeleteSuccessContent />
     </ProtectedRoute>
   );
 }
