+++ bods-nextjs/app/publish/org/[orgId]/dataset/avl/[datasetId]/success/page.tsx
@@ -1,5 +1,10 @@
+import type { Metadata } from 'next';
 import { ProtectedRoute } from '@/components/auth/ProtectedRoute';
 import { AvlSuccessPanel } from '../../_components/AvlSuccessPanel';
+
+export const metadata: Metadata = {
+  title: 'Data feed published',
+};
 
 export default function AVLPublishSuccessPage() {
   return (
