+++ bods-nextjs/app/publish/org/[orgId]/dataset/avl/[datasetId]/update/success/page.tsx
@@ -1,5 +1,10 @@
+import type { Metadata } from 'next';
 import { ProtectedRoute } from '@/components/auth/ProtectedRoute';
 import { AvlSuccessPanel } from '../../../_components/AvlSuccessPanel';
+
+export const metadata: Metadata = {
+  title: 'Data feed updated',
+};
 
 export default function AVLUpdateSuccessPage() {
   return (
