+++ bods-nextjs/app/publish/org/[orgId]/dataset/avl/_components/AvlSuccessPanel.tsx
@@ -2,7 +2,6 @@
 
 import Link from 'next/link';
 import { useParams } from 'next/navigation';
-import { useEffect } from 'react';
 
 type AvlSuccessPanelProps = {
   update: boolean;
@@ -14,10 +13,6 @@ export function AvlSuccessPanel({ update }: AvlSuccessPanelProps) {
   const datasetId = params.datasetId as string;
   const listUrl = `/publish/org/${orgId}/dataset/avl`;
   const detailUrl = `/publish/org/${orgId}/dataset/avl/${datasetId}`;
-
-  useEffect(() => {
-    document.title = update ? 'Data feed updated' : 'Data feed published';
-  }, [update]);
 
   return (
     <div className="govuk-width-container">
